Fix cross-implementation inconsistencies
Sweep every implementation for drift from IMPLEMENTATION.md and from each
other, then close the gaps: C's _f, Fortran's add_cell return, Pony's inline
world size, the prepopulate_neighbours bounds checks in F#/Lua/OCaml/Scala,
the existing local in add_cell, Erlang's and OCaml's exception messages,
self-derived infinity sentinels in AssemblyScript/Navi/Gleam/Haskell, missing
render and alive_neighbours variants in C/Fortran/Odin/Ballerina, comment
wording, constant placement, DIRECTIONS layout in C/Rust/Zig, and the README
Notes sections. Every "slower"/"fastest" label is measured, and each commented
variant was activated in a scratch copy and compiled. Also fixes the
pre-existing V build break, where time.sys_mono_now() now returns u64.

+++ v/play.v
@@ -28,7 +28,7 @@ fn run() {
     tick_start := time.sys_mono_now()
     world.tick()
     tick_finish := time.sys_mono_now()
-    tick_time := (tick_finish - tick_start)
+    tick_time := f64(tick_finish - tick_start)
     total_tick += tick_time
     lowest_tick = math.min(lowest_tick, tick_time)
     avg_tick := (total_tick / world.tick)
@@ -36,7 +36,7 @@ fn run() {
     render_start := time.sys_mono_now()
     rendered := world.render()
     render_finish := time.sys_mono_now()
-    render_time := (render_finish - render_start)
+    render_time := f64(render_finish - render_start)
     total_render += render_time
     lowest_render = math.min(lowest_render, render_time)
     avg_render := (total_render / world.tick)
